fix more warnings

+++ src/c/ncm_song.h
@@ -78,7 +78,7 @@ int32 ncm_song_format_numeric_tag(char *buffer, int32 buffer_cap,
                                   char *tag, int32 tag_len);
 int32 ncm_song_format_track_number(char *buffer, int32 buffer_cap,
                                    char *tag, int32 tag_len);
-int32 ncm_song_show_time(int32 length, char *buffer, int32 buffer_cap);
+int32 ncm_song_show_time(int64 length, char *buffer, int32 buffer_cap);
 NcmBuffer ncm_song_getter_buffer(NcmSong *song,
                                  enum NcmSongGetter getter, int32 idx);
 NcmBuffer ncm_song_tags_buffer(NcmSong *song, enum NcmSongGetter getter,
